charge_pump: modify the chip id output format
charge_pump:
Modify the chip id output format.

Reviewed-on: https://gerrit.mot.com/3169432
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Submit-Approved: Jira Key

+++ drivers/power/mmi_glink_charger/device_glink/charge_pump_glink.h
@@ -20,7 +20,7 @@ struct charge_pump_dev_info
 	bool ovpgate;
 	bool manual;
 	bool otg_en;
-	u16 chip_id;
+	s32 chip_id;
 	u8 chg_role;
 	u8 work_mode;
 	u8 int_stat;
